// src/generated/model/studentEnrollmentResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface StudentEnrollmentResponse {
  id?: string;
  studentId?: string;
  academicYearId?: string;
  intakeId?: string;
  batchId?: string;
  programId?: string;
  programLevelId?: string;
  classId?: string;
  enrollmentType?: string;
  status?: string;
  enrollmentDate?: string;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}